refactor: add changes to role manager service
