+++ domains/middleware/authorization.go
@@ -45,7 +45,7 @@ func NewAuthorization(entityType string, svc domains.Service, authz smqauthz.Aut
 		return nil, err
 	}
 
-	ram, err := rmMW.NewRoleManagerAuthorizationMiddleware(entityType, svc, authz, rolesOpPerm, callout)
+	ram, err := rmMW.NewAuthorization(entityType, svc, authz, rolesOpPerm, callout)
 	if err != nil {
 		return nil, err
 	}
