+++ domains/middleware/metrics.go
@@ -27,7 +27,7 @@ type metricsMiddleware struct {
 
 // NewMetrics instruments core service by tracking request count and latency.
 func NewMetrics(svc domains.Service, counter metrics.Counter, latency metrics.Histogram) domains.Service {
-	rmmw := rmMW.NewRoleManagerMetricsMiddleware("domains", svc, counter, latency)
+	rmmw := rmMW.NewMetrics("domains", svc, counter, latency)
 
 	return &metricsMiddleware{
 		counter:                      counter,
