--- pkg/roles/rolemanager/doc.go
@@ -1,4 +1,0 @@
-// Copyright (c) Abstract Machines
-// SPDX-License-Identifier: Apache-2.0
-
-package rolemanager
