+++ pkg/roles/rolemanager/middleware/doc.go
@@ -1,0 +1,9 @@
+// Copyright (c) Abstract Machines
+// SPDX-License-Identifier: Apache-2.0
+
+// Package middleware provides authorization, logging, metrics and tracing middleware
+// for SuperMQ RoleManager service.
+//
+// For more details about tracing instrumentation for SuperMQ refer to the
+// documentation at https://docs.supermq.abstractmachines.fr/tracing/.
+package middleware
