+++ pkg/roles/rolemanager/middleware/logging.go
@@ -23,8 +23,8 @@ type RoleManagerLoggingMiddleware struct {
 	logger  *slog.Logger
 }
 
-// NewRoleManagerLoggingMiddleware adds logging facilities to the core service.
-func NewRoleManagerLoggingMiddleware(svcName string, svc roles.RoleManager, logger *slog.Logger) RoleManagerLoggingMiddleware {
+// NewLogging adds logging facilities to the core service.
+func NewLogging(svcName string, svc roles.RoleManager, logger *slog.Logger) RoleManagerLoggingMiddleware {
 	return RoleManagerLoggingMiddleware{
 		svcName: svcName,
 		svc:     svc,
